+++ src/bindings/bnd_embedded_file.cpp
@@ -17,7 +17,6 @@ BND_File3dmEmbeddedFile::BND_File3dmEmbeddedFile(ON_EmbeddedFile* ef, const ON_M
   SetTrackedPointer(ef, compref);
 }
 
-//works for python
 BND_File3dmEmbeddedFile* BND_File3dmEmbeddedFile::Read(const std::wstring& f) // Static.
 {
   auto* ef = new ON_EmbeddedFile;
@@ -122,7 +121,7 @@ void initEmbeddedFileBindings(pybind11::module& m)
     .def(py::init<const BND_File3dmEmbeddedFile&>(), py::arg("other"))
     .def_static("Read", &BND_File3dmEmbeddedFile::Read, py::arg("fileName"))
     .def_property_readonly("Length", &BND_File3dmEmbeddedFile::GetLength)
-    .def_property_readonly("FileName", &BND_File3dmEmbeddedFile::GetFilename)
+    .def_property("Filename", &BND_File3dmEmbeddedFile::GetFilename,  &BND_File3dmEmbeddedFile::SetFilename)
     .def("Write", &BND_File3dmEmbeddedFile::Write, py::arg("fileName"))
     .def("Clear", &BND_File3dmEmbeddedFile::Clear)
     ;
@@ -145,7 +144,7 @@ void initEmbeddedFileBindings(void*)
     //.class_function("read", &BND_File3dmEmbeddedFile::Read, allow_raw_pointers())
     .class_function("fromByteArray", &BND_File3dmEmbeddedFile::WasmFromByteArray, allow_raw_pointers())
     .property("length", &BND_File3dmEmbeddedFile::GetLength)
-    .property("fileName", &BND_File3dmEmbeddedFile::GetFilename)
+    .property("fileName", &BND_File3dmEmbeddedFile::GetFilename, &BND_File3dmEmbeddedFile::SetFilename)
     //.function("setFileName", &BND_File3dmEmbeddedFile::SetFilename) //TODO
     .function("write", &BND_File3dmEmbeddedFile::Write, allow_raw_pointers()) //should return some sort of buffer that can be saved with the FileAPI
     .function("clear", &BND_File3dmEmbeddedFile::Clear)